changed back to 32 bit floats

+++ Demo2/Computer_Vision/Demo2.py
@@ -173,9 +173,9 @@ def send_instructions():
        #just three bytes
        instruction_array = [instructions["good_angle"], instructions["good_distance"], instructions["arrow"]]
        #special handling for floats which will require an extra 8 bytes 4 bytes for each float
-       angle_in_bytes = list(struct.pack('f', np.float16(instructions["angle"])))
+       angle_in_bytes = list(struct.pack('f', instructions["angle"]))
        instruction_array += angle_in_bytes
-       distance_in_bytes = list(struct.pack('f', np.float16(instructions["distance"])))
+       distance_in_bytes = list(struct.pack('f',instructions["distance"]))
        instruction_array += distance_in_bytes
        #print(instruction_array)
        #total send is 3 + 8 bytes = 11 bytes, much better than the 24 from the previous implementation
